+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Android Studio 3.1.2 に対応した
 - Kotlin を 1.2.41 に上げた
 - connect オプションの spotlight に対応した
+- PeerConnectionFactory を builder から作るよう修正した
 
 ### ADD